fixed up issues with tests refactor
