+++ arcana/cli/tests/test_store_cli.py
@@ -17,8 +17,8 @@ def test_store_cli(xnat_repository, cli_runner, work_dir):
         # List all saved and built-in stores
         result = cli_runner(ls, [])
         assert result.exit_code == 0
-        assert 'bids - arcana.data.stores.bids:BidsFormat' in result.output
-        assert 'file - arcana.data.stores.common:FileSystem' in result.output
+        assert 'bids - arcana.data.stores.bids.structure:BidsFormat' in result.output
+        assert 'file - arcana.data.stores.common.file_system:FileSystem' in result.output
         assert 'test-xnat - arcana.data.stores.medimage.xnat.api:Xnat' in result.output
         assert '    server: ' + xnat_repository.server in result.output
