+++ arcana/data/stores/bids/__init__.py
@@ -1,0 +1,2 @@
+from .structure import BidsFormat
+from .dataset import BidsDataset
