+++ arcana/tests/fixtures/common.py
@@ -19,6 +19,17 @@ from pathlib import Path
 import pytest
 import attr
 from pydra import mark
+from click.testing import CliRunner
+
+
+@pytest.fixture
+def work_dir():
+    # work_dir = Path.home() / '.arcana-tests'
+    # work_dir.mkdir(exist_ok=True)
+    # return work_dir
+    work_dir = mkdtemp()
+    yield Path(work_dir)
+    shutil.rmtree(work_dir)
 
 
 @mark.task
@@ -372,3 +383,29 @@ def create_test_file(fname, dpath):
         (dpath / fpath).unlink()
         fpath = Path(fname)
     return fpath
+
+
+# For debugging in IDE's don't catch raised exceptions and let the IDE
+# break at it
+if os.getenv('_PYTEST_RAISE', "0") != "0":
+
+    @pytest.hookimpl(tryfirst=True)
+    def pytest_exception_interact(call):
+        raise call.excinfo.value
+
+    @pytest.hookimpl(tryfirst=True)
+    def pytest_internalerror(excinfo):
+        raise excinfo.value
+
+    catch_cli_exceptions = False
+else:
+    catch_cli_exceptions = True
+
+@pytest.fixture
+def cli_runner():
+    def invoke(*args, **kwargs):
+        runner = CliRunner()
+        result = runner.invoke(*args, catch_exceptions=catch_cli_exceptions,
+                               **kwargs)
+        return result
+    return invoke
